add: env creation scripts
add: set guest storage target's runsv init scripts
add: setup template for iso9660 generation from the virtual disk
add: submodules for apk-tools and ebpf_exporter
add: create scripts for virtual raw disk population with qcow2

+++ scripts/firecracker-startup.sh
@@ -1,6 +1,18 @@
 #!/bin/sh
 #
+build_fck() {
 
+if [ -z "$(ls -l ./assets/firecracker)" ]; then
+    git clone https://github.com/firecracker-microvm/firecracker ./assets
+    cd firecracker || return
+    tools/devtool build
+    toolchain="$(uname -m)-unknown-linux-musl"
+
+    cd - || return
+fi
+}
+
+runsv_service() {
 mkdir -p /etc/runit/sv/firecracker
 
 cat <<"EOF" > /etc/runit/sv/firecracker/fck-up.sh
@@ -10,5 +22,5 @@ firecracker -p
 EOF
 
 ln -s /etc/runit/sv/firecracker/fck-up.sh /run/runit/service/
-
+}
